move secret property tests to common file
